pyproject.toml
[project]
name = "langgraph_practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=0.3.77",
    "langchain-openai>=0.3.34",
    "langgraph>=0.6.8",
    "python-dotenv>=1.1.1",
    "typing-extensions>=4.15.0",
]
